refactor: update dtos path
